add authorization in swagger setting and example
