+++ lib/index.js
@@ -7,33 +7,31 @@
 
 'use strict';
 
+const utils = require('./utils');
+
 module.exports = function (config) {
     return function () {
         var app = this;
         var services = {};
-        config = config || {};
 
         // Enable the swagger Plugin
         app.enable('feathers swagger');
 
         // Apply configuration
-        var rootDoc = {};
-        var basePath = config.basePath || '';
-        var docsPath = config.docsPath || '/docs';
-        var docExt = config.docExt || '';
+        var rootDoc = Object.assign({
+            paths: {},
+            definitions: {},
+            swagger: '2.0',
+            schemes: ['http'],
+            tags: [],
+            basePath: '',
+            docsPath: '/docs',
+            consumes: ['application/json'],
+            produces: ['application/json'],
+            info: {}
+        }, config || {});
 
-        // Setup docs from config
-        rootDoc.info = config.info || {};
-
-        // by 2.0
-        rootDoc.paths = config.paths || {};
-        rootDoc.definitions = config.definitions || {};
-        rootDoc.swagger = config.swagger || '2.0';
-        rootDoc.schemes = ['http'];
-        rootDoc.tags = [];
-        rootDoc.basePath = basePath;
-        rootDoc.consumes = ['application/json'];
-        rootDoc.produces = ['application/json'];
+        var docsPath = rootDoc.docsPath;
 
         // Create API for Documentation
         app.get(docsPath, function (req, res) {
@@ -42,48 +40,29 @@ module.exports = function (config) {
 
         // Optional: Register this plugin as a Feathers provider
         app.providers.push(function (path, service) {
-            // console.log(path, service);
             services[path] = service;
             service.docs = service.docs || {};
             // Load documentation from service, if available.
             var doc = service.docs;
 
             // by 2.0
-            var pathObj = rootDoc.paths || {},
+            var pathObj = rootDoc.paths,
                 withIdKey = '/'+path+'/{resourceId}',
                 withoutIdKey = '/'+path;
 
             if(typeof pathObj[withoutIdKey] === 'undefined'){ pathObj[withoutIdKey] = {}; }
             if(typeof pathObj[withIdKey] === 'undefined'){ pathObj[withIdKey] = {}; }
-            
+
             if(typeof doc.definition !== 'undefined'){
                 rootDoc.definitions[path] = doc.definition;
             }
             if(typeof doc.definitions !== 'undefined'){
-                rootDoc.definitions = Object.assign(rootDoc.definitions,doc.definitions);
-            }
-
-            function Operation(method, service, defaults) {
-                defaults = defaults || {};
-                // Find is available
-                var operation = service.docs[method] || {};
-                operation.parameters = operation.parameters || defaults.parameters || [];
-                operation.responses = operation.responses || defaults.responses || [];
-                operation.description = operation.description || defaults.description || '';
-                operation.summary = operation.summary || defaults.summary || '';
-                operation.tags = operation.tags || defaults.tags || [];
-                operation.consumes = operation.consumes || defaults.consumes || [];
-                operation.produces = operation.produces || defaults.produces || [];
-                operation.security = operation.security || defaults.security || [];
-                operation.securityDefinitions = operation.securityDefinitions || defaults.securityDefinitions || [];
-                // Clean up
-                delete service.docs[method]; // Remove `find` from `docs`
-                return operation;
+                rootDoc.definitions = Object.assign(rootDoc.definitions, doc.definitions);
             }
 
             // FIND
             if (typeof service.find === 'function') {
-                pathObj[withoutIdKey].get = new Operation('find', service, {
+                pathObj[withoutIdKey].get = new utils.Operation('find', service, {
                     tags: [path],
                     description: 'Retrieves a list of all resources from the service.',
                     produces: rootDoc.produces,
@@ -92,7 +71,7 @@ module.exports = function (config) {
             }
             // GET
             if (typeof service.get === 'function') {
-                pathObj[withIdKey].get = new Operation('get', service, {
+                pathObj[withIdKey].get = new utils.Operation('get', service, {
                     tags: [path],
                     description: 'Retrieves a single resource with the given id from the service.',
                     parameters: [{
@@ -104,7 +83,7 @@ module.exports = function (config) {
                     }],
                     responses: {
                         '200': {
-                            description: 'successful operation',
+                            description: 'successful utils.operation',
                             schema: {
                                 '$ref': '#/definitions/'+path
                             }
@@ -116,7 +95,7 @@ module.exports = function (config) {
             }
             // CREATE
             if (typeof service.create === 'function') {
-                pathObj[withoutIdKey].post = new Operation('create', service, {
+                pathObj[withoutIdKey].post = new utils.Operation('create', service, {
                     tags: [path],
                     description: 'Creates a new resource with data.',
                     parameters: [{
@@ -131,7 +110,7 @@ module.exports = function (config) {
             }
             // UPDATE
             if (typeof service.update === 'function') {
-                pathObj[withIdKey].put = new Operation('update', service, {
+                pathObj[withIdKey].put = new utils.Operation('update', service, {
                     tags: [path],
                     description: 'Updates the resource identified by id using data.',
                     parameters: [{
@@ -152,7 +131,7 @@ module.exports = function (config) {
             }
             // REMOVE
             if (typeof service.remove === 'function') {
-                pathObj[withIdKey].delete = new Operation('remove', service, {
+                pathObj[withIdKey].delete = new utils.Operation('remove', service, {
                     tags: [path],
                     description: 'Removes the resource with id.',
                     parameters: [{
@@ -167,108 +146,11 @@ module.exports = function (config) {
                 });
             }
 
-            function Tag(name, options){
-                options = options || {};
-                var result = {};
-
-                result.name = name;
-                result.description = options.description || 'Operations about this resource.';
-                result.externalDocs = options.externalDocs || {};
-
-                return result;
-            }
-
             rootDoc.paths = pathObj;
-            rootDoc.tags.push(new Tag(path, doc));
-
-            // Create handler for serving the service's documentation
-            app.get(docsPath + '/' + path + docExt, function (req, res) {
-                res.json(doc);
-            });
-
+            rootDoc.tags.push(new utils.Tag(path, doc));
         });
 
     };
 };
 
-module.exports.util = {
-    Definition: function(model, type, properties){
-        type = type ? type : 'object';
-        properties = properties ? properties : {};
-        var result = {
-                type: type,
-                properties: {}
-            },
-            keys = typeof model.attributes !== 'undefined' ? Object.keys(model.attributes) : [];
-        
-        for (var i = 0; i < keys.length; i++) {
-            var attrName = keys[i],
-                attrType = model.attributes[attrName].type.constructor.prototype.key,
-                propertie = new module.exports.util.Propertie(attrType);
-            result.properties[attrName] = propertie;
-        }
-        result.properties = Object.assign(result.properties, properties);
-
-        return result;
-    },
-    Propertie: function(type, items){
-        items = items ? items : {};
-        var result = {
-            type: module.exports.util.getType(type),
-            format: module.exports.util.getFormat(type)
-        };
-
-        if(type === 'ARRAY'){
-            result.items = items;
-        }
-
-        return result;
-    },
-    getType: function(type){
-        switch(type){
-            case 'STRING':
-            case 'CHAR':
-            case 'TEXT':
-            case 'BLOB':
-            case 'DATE':
-            case 'DATEONLY':
-            case 'TIME':
-            case 'NOW':
-                return 'string';
-            case 'INTEGER':
-            case 'BIGINT':
-                return 'integer';
-            case 'FLOAT':
-            case 'DOUBLE':
-            case 'DECIMAL':
-                return 'number';
-            case 'BOOLEAN':
-                return 'boolean';
-            case 'ARRAY':
-                return 'array';
-            default:
-                return '';
-        }
-    },
-    getFormat: function(type){
-        switch(type){
-            case 'INTEGER':
-            case 'DECIMAL':
-                return 'int32';
-            case 'BIGINT':
-                return 'int64';
-            case 'FLOAT':
-                return 'float';
-            case 'DOUBLE':
-                return 'double';
-            case 'DATE':
-            case 'DATEONLY':
-                return 'date';
-            case 'TIME':
-            case 'NOW':
-                return 'date-time';
-            default:
-                return '';
-        }
-    }
-};
+module.exports.utils = utils;
